Added consts and rename parse
