+++ src/kraken/std/git/tasks/const.py
@@ -1,0 +1,26 @@
+GITIGNORE_TASK_NAME = "gitignore"
+
+DEFAULT_GITIGNORE_TOKENS = [
+    # Platforms
+    "macos",
+    "linux",
+    "windows",
+    # IDEs
+    "visualstudiocode",
+    "vim",
+    "emacs",
+    "clion",
+    "intellij",
+    "pycharm",
+    "jupyternotebooks",
+    # Tooling
+    "git",
+    "gcov",
+    "node",
+    "yarn",
+    # Languages
+    "python",
+    "rust",
+    "react",
+    "matlab",
+]
